# api/src/handlers/wallet_auth.rs
use super::*;

// ============================================================================
// Wallet Authentication Endpoints
// ============================================================================

/// Status a newly registered (not yet approved) account is created with.
///
/// Must stay one of `state::USER_STATUSES`, and must NOT be `"active"` —
/// registration is an Admin-approved flow, not self-service activation.
const REGISTERED_USER_STATUS: &str = "pending";

/// Longest staff contact number accepted.
///
/// Generous enough for `+27 (0)11 555 0100 x4821` and short enough that the
/// field cannot be used as free-text storage. The sealed blob has no column
/// width of its own, so this is the only bound there is.
pub(crate) const MAX_STAFF_PHONE_LEN: usize = 32;

/// Trim a submitted staff phone number, or explain why it was refused.
///
/// An empty or whitespace-only string becomes `None`: a field the
/// administrator left blank is absent, not a contact number that happens to be
/// "" (CLAUDE.md rule 9). The number itself is not pattern-validated — MediChain
/// spans several national dialling plans, and a format rule written against one
/// of them would reject valid numbers from the others.
fn normalise_staff_phone(raw: Option<&str>) -> Result<Option<String>, HttpResponse> {
    let trimmed = raw.map(str::trim).filter(|value| !value.is_empty());
    if trimmed.is_some_and(|value| value.chars().count() > MAX_STAFF_PHONE_LEN) {
        return Err(HttpResponse::BadRequest().json(ErrorResponse {
            error: format!("Phone number must be at most {MAX_STAFF_PHONE_LEN} characters"),
            code: "INVALID_PHONE".to_string(),
        }));
    }
    Ok(trimmed.map(str::to_string))
}

/// Bootstrap request - for creating first admin
#[derive(Debug, Deserialize)]
pub struct BootstrapAdminRequest {
    pub wallet_address: String,
    pub name: String,
    pub username: Option<String>,
    pub secret_key: String, // Environment variable MEDICHAIN_BOOTSTRAP_KEY must match
}

/// Bootstrap response
#[derive(Debug, Serialize)]
pub struct BootstrapAdminResponse {
    pub success: bool,
    pub admin: WalletUserInfo,
    pub message: String,
}

/// Bootstrap first admin (only works when no users exist)
/// This endpoint allows the first admin to be created without authentication
/// SECURITY: Requires MEDICHAIN_BOOTSTRAP_KEY environment variable to match
/// In production, this key MUST be set via environment variable
#[post("/api/auth/bootstrap")]
pub async fn bootstrap_admin(
    data: web::Data<AppState>,
    body: web::Json<BootstrapAdminRequest>,
) -> impl Responder {
    // Check if running in demo mode
    let is_demo = std::env::var("IS_DEMO")
        .map(|v| v == "true" || v == "1")
        .unwrap_or(false);

    // Check bootstrap key from environment
    // SECURITY: In production (non-demo), require explicit key from environment
    let bootstrap_key = match std::env::var("MEDICHAIN_BOOTSTRAP_KEY") {
        Ok(key) => key,
        Err(_) if is_demo => "medichain-dev-bootstrap-2024".to_string(),
        Err(_) => {
            return HttpResponse::Forbidden().json(ErrorResponse {
                error: "MEDICHAIN_BOOTSTRAP_KEY environment variable required in production"
                    .to_string(),
                code: "MISSING_BOOTSTRAP_KEY".to_string(),
            });
        }
    };

    if body.secret_key != bootstrap_key {
        return HttpResponse::Forbidden().json(ErrorResponse {
            error: "Invalid bootstrap key".to_string(),
            code: "INVALID_BOOTSTRAP_KEY".to_string(),
        });
    }

    // Include inactive users that are intentionally absent from the auth cache.
    match data.has_any_persisted_user().await {
        Ok(true) => {
            return HttpResponse::Conflict().json(ErrorResponse {
                error: "Bootstrap not allowed - users already exist. Use /api/auth/register with admin credentials.".to_string(),
                code: "BOOTSTRAP_NOT_ALLOWED".to_string(),
            });
        }
        Ok(false) => {}
        Err(e) => {
            log::error!("Bootstrap user existence check failed: {}", e);
            return HttpResponse::ServiceUnavailable().json(ErrorResponse {
                error: "User storage is unavailable".to_string(),
                code: "USER_PERSISTENCE_UNAVAILABLE".to_string(),
            });
        }
    }

    // Validate wallet address format
    if !is_valid_wallet_address(&body.wallet_address) {
        return HttpResponse::BadRequest().json(ErrorResponse {
            error:
                "Invalid wallet address format. Must be SS58 encoded (starts with 5, 45-50 chars)"
                    .to_string(),
            code: "INVALID_WALLET_ADDRESS".to_string(),
        });
    }

    // Create first admin
    let admin = User {
        wallet_address: body.wallet_address.clone(),
        username: body.username.clone(),
        name: body.name.clone(),
        role: Role::Admin,
        created_at: Utc::now(),
        created_by: None, // Self-created
        linked_patient_id: None,
        email: None,
        phone: None,
        department: None,
        specialty: None,
        license_number: None,
        status: "active".to_string(),
        last_login: None,
    };

    if let Err(e) = data.persist_then_cache_user(admin.clone()).await {
        log::error!("Failed to persist bootstrap administrator: {}", e);
        return HttpResponse::ServiceUnavailable().json(ErrorResponse {
            error: "Administrator account could not be persisted".to_string(),
            code: "USER_PERSISTENCE_UNAVAILABLE".to_string(),
        });
    }

    log::info!(
        "Bootstrap: First admin created - wallet={}, name={}",
        body.wallet_address,
        body.name
    );

    HttpResponse::Created().json(BootstrapAdminResponse {
        success: true,
        admin: WalletUserInfo {
            wallet_address: body.wallet_address.clone(),
            name: body.name.clone(),
            role: "Admin".to_string(),
            username: body.username.clone(),
            linked_patient_id: None,
        },
        message: "First admin created successfully. System is now bootstrapped.".to_string(),
    })
}

/// Register a new user with wallet address (Admin only)
/// This creates a new user account linked to a blockchain wallet
#[post("/api/auth/register")]
pub async fn wallet_register(
    data: web::Data<AppState>,
    req: HttpRequest,
    body: web::Json<WalletRegisterRequest>,
) -> impl Responder {
    // Get current user (must be admin to register new users)
    let current_user_id = match get_current_user_id(&req) {
        Some(id) => id,
        None => {
            return HttpResponse::Unauthorized().json(ErrorResponse {
                error: "Missing X-User-Id header".to_string(),
                code: "UNAUTHORIZED".to_string(),
            });
        }
    };

    let current_user = match get_user(&data, &current_user_id) {
        Some(u) => u,
        None => {
            return HttpResponse::Unauthorized().json(ErrorResponse {
                error: "Admin user not found".to_string(),
                code: "USER_NOT_FOUND".to_string(),
            });
        }
    };

    // Only admin can register new users
    if !current_user.role.is_admin() {
        return HttpResponse::Forbidden().json(ErrorResponse {
            error: "Only Admin can register new users".to_string(),
            code: "INSUFFICIENT_ROLE".to_string(),
        });
    }

    // Validate wallet address format
    if !is_valid_wallet_address(&body.wallet_address) {
        return HttpResponse::BadRequest().json(ErrorResponse {
            error:
                "Invalid wallet address format. Must be SS58 encoded (starts with 5, 45-50 chars)"
                    .to_string(),
            code: "INVALID_WALLET_ADDRESS".to_string(),
        });
    }

    // Check if wallet already registered
    {
        let users = data.users.read().unwrap();
        if users.contains_key(&body.wallet_address) {
            return HttpResponse::Conflict().json(ErrorResponse {
                error: "Wallet address already registered".to_string(),
                code: "WALLET_ALREADY_REGISTERED".to_string(),
            });
        }
    }

    // Parse role
    let role = match parse_role(&body.role) {
        Ok(r) => r,
        Err(e) => {
            return HttpResponse::BadRequest().json(ErrorResponse {
                error: e,
                code: "INVALID_ROLE".to_string(),
            });
        }
    };

    // Cannot register Admin role
    if role.is_admin() {
        return HttpResponse::Forbidden().json(ErrorResponse {
            error: "Cannot register Admin role via API".to_string(),
            code: "CANNOT_REGISTER_ADMIN".to_string(),
        });
    }

    // Accepted since migration `20260910000007`, which gave staff contact
    // details an encrypted home. `persist_user` seals this into
    // `user_profiles.contact_encrypted`; the plaintext `phone` column stays
    // deprecated and unwritten. Before that column existed this handler refused
    // any non-empty phone outright, which left an administrator unable to
    // record a way to contact the clinician they had just onboarded.
    let phone = match normalise_staff_phone(body.phone.as_deref()) {
        Ok(phone) => phone,
        Err(response) => return response,
    };

    // Create new user
    let user = User {
        wallet_address: body.wallet_address.clone(),
        username: body.username.clone(),
        name: body.name.clone(),
        role: role.clone(),
        created_at: Utc::now(),
        created_by: Some(current_user_id.clone()),
        linked_patient_id: None,
        email: body.email.clone(),
        phone,
        department: body.department.clone(),
        specialty: body.specialty.clone(),
        license_number: body.license_number.clone(),
        // Deliberate approval gate: `get_user` resolves only `active` users, so
        // this account stays unusable until an Admin activates it. The response
        // below says so explicitly.
        status: REGISTERED_USER_STATUS.to_string(),
        last_login: None,
    };

    if let Err(e) = data.persist_then_cache_user(user).await {
        log::error!("Failed to persist new user: {e}");
        return HttpResponse::ServiceUnavailable().json(ErrorResponse {
            error: "User registration could not be persisted".to_string(),
            code: "USER_PERSISTENCE_UNAVAILABLE".to_string(),
        });
    }

    log::info!(
        "New user registered: wallet={}, name={}, role={} by admin={}",
        body.wallet_address,
        body.name,
        role,
        current_user_id
    );

    HttpResponse::Created().json(WalletRegisterResponse {
        success: true,
        wallet_address: body.wallet_address.clone(),
        role: role.to_string(),
        status: REGISTERED_USER_STATUS.to_string(),
        activation_required: true,
        message: "User registered. An Admin must set status to \"active\" via PUT /api/users/{wallet_address} before the account can sign in."
            .to_string(),
    })
}