# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MediChain API
  description: |
    MediChain REST API provides secure access to patient identity, medical records, 
    and access control functionality for a blockchain-based national health ID system.
    
    © 2025 Lukau Invasion (Pty) Ltd. All rights reserved.
    
    ## Authentication
    
    MediChain uses **wallet-based blockchain authentication** with SS58 addresses. 
    For API testing, include the `X-User-Id` header with a valid wallet address.
    
    ## Role-Based Access Control
    
    All endpoints implement RBAC with the following roles:
    - **Admin** - Full system access, user management
    - **Doctor** - Patient registration, medical record access/edit
    - **Nurse** - Patient registration, medical record access/edit
    - **LabTechnician** - Lab results upload, limited record access
    - **Pharmacist** - Prescription verification, medication records
    - **Patient** - Read-only access to own records
    
    ## Pagination
    
    List endpoints support pagination via query parameters:
    - `page` - Page number (1-indexed, default: 1)
    - `limit` - Items per page (default: 20, max: 100)
    
  version: "1.0.0"
  contact:
    name: MediChain Support
    email: support@medichain.ng
  license:
    name: Proprietary
    url: https://medichain.ng/license
servers:
  - url: http://localhost:8090
    description: Local development server
  - url: https://api.medichain.ng
    description: Production server
...
security:
  - WalletAuth: []
paths:
...
  /api/auth/bootstrap:
    post:
      tags: [Authentication]
      summary: Bootstrap admin account
      description: |
        Creates the initial admin account. Only works once and requires 
        BOOTSTRAP_KEY environment variable in production.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BootstrapRequest'
      responses:
        '201':
          description: Admin account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BootstrapResponse'
        '400':
          description: Admin already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Register new user
      description: Register a new user with their wallet address. Only Admin can register healthcare providers.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalletRegisterRequest'
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletRegisterResponse'
        '400':
          description: Invalid request
        '403':
          description: Insufficient permissions
...
components:
  securitySchemes:
    WalletAuth:
      type: apiKey
      in: header
      name: X-User-Id
      description: SS58 encoded wallet address
...
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        code:
          type: string
...
    WalletUserInfo:
      type: object
      properties:
        wallet_address:
          type: string
        name:
          type: string
        role:
          type: string
        username:
          type: string
        linked_patient_id:
          type: string
    WalletRegisterRequest:
      type: object
      required: [wallet_address, name, role]
      properties:
        wallet_address:
          type: string
        name:
          type: string
        username:
          type: string
        role:
          type: string
          enum: [Admin, Doctor, Nurse, LabTechnician, Pharmacist, Patient]
    WalletRegisterResponse:
      type: object
      properties:
        success:
          type: boolean
        user:
          $ref: '#/components/schemas/WalletUserInfo'
        message:
          type: string
...
    BootstrapRequest:
      type: object
      required: [wallet_address, name]
      properties:
        wallet_address:
          type: string
        name:
          type: string
        bootstrap_key:
          type: string
          description: Required in production
    BootstrapResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        admin:
          $ref: '#/components/schemas/WalletUserInfo'
...